requirements.txt
pandas
numpy
statsmodels
# openpyxl # 如果你的数据源是Excel 
python-dotenv 